Deprecate old Obj export method. Implement ObjOptions in exporter.

+++ probuilder2.0/Assets/ProCore/ProBuilder/Editor/EditorCore/MenuActions/Export/ExportObj.cs
@@ -4,6 +4,7 @@ using ProBuilder2.Common;
 using ProBuilder2.EditorCommon;
 using ProBuilder2.Interface;
 using System.Linq;
+using System.IO;
 using System.Collections.Generic;
 using Parabox.STL;
 
@@ -39,17 +40,12 @@ namespace ProBuilder2.Actions
 				return new pb_ActionResult(Status.Success, "Export OBJ");
 		}
 
-		public static string ExportWithFileDialog(IEnumerable<pb_Object> meshes)
+		/**
+		 *	Prompt user for a save file location and export meshes as Obj.
+		 */
+		public static string ExportWithFileDialog(IEnumerable<pb_Object> meshes, bool asGroup = true, pb_ObjOptions options = null)
 		{
-			pb_Object first = meshes.FirstOrDefault();
-
-			string name = first != null ? first.name : null;
-			if(string.IsNullOrEmpty(name))
-				return "";
-
-			string path = EditorUtility.SaveFolderPanel("Export to Obj", "Assets", name);
-
-			if(string.IsNullOrEmpty(path))
+			if(meshes == null || meshes.Count() < 1)
 				return null;
 
 			IEnumerable<pb_Model> models = meshes.Select(x =>
@@ -58,10 +54,58 @@ namespace ProBuilder2.Actions
 					x.GetComponent<MeshRenderer>().sharedMaterials,
 					x.transform.localToWorldMatrix));
 
-			string obj, mat;
-			pb_Obj.Export(name, models, out obj, out mat);
-			pb_FileUtil.WriteFile(string.Format("{0}/{1}.obj", path, name), obj);
-			pb_FileUtil.WriteFile(string.Format("{0}/{1}.mtl", path, name), mat);
+			string path = null;
+
+			if(asGroup || models.Count() < 2)
+			{
+				pb_Object first = meshes.FirstOrDefault();
+				string name = first != null ? first.name : "ProBuilderModel";
+				path = EditorUtility.SaveFilePanel("Export to Obj", "Assets", name, "obj");
+
+				if(string.IsNullOrEmpty(path))
+					return null;
+
+				string obj, mat;
+
+				pb_Obj.Export(name, models, out obj, out mat, options);
+
+				try
+				{
+					path = Path.GetDirectoryName(path);
+					pb_FileUtil.WriteFile(string.Format("{0}/{1}.obj", path, name), obj);
+					// Spaces are not allowed in mtlib name
+					pb_FileUtil.WriteFile(string.Format("{0}/{1}.mtl", path, name.Replace(" ", "_")), mat);
+				}
+				catch(System.Exception e)
+				{
+					Debug.LogWarning(string.Format("Failed writing obj to path: {0}\n{1}", string.Format("{0}/{1}.obj", path, name), e.ToString()));
+				}
+			}
+			else
+			{
+				path = EditorUtility.SaveFolderPanel("Export to Obj", "Assets", "");
+
+				if(string.IsNullOrEmpty(path) || !Directory.Exists(path))
+					return null;
+
+				foreach(pb_Model model in models)
+				{
+					string name = model.name;
+					string obj, mat;
+
+					pb_Obj.Export(name, new List<pb_Model>() { model }, out obj, out mat, options);
+
+					try
+					{
+						pb_FileUtil.WriteFile(string.Format("{0}/{1}.obj", path, name), obj);
+						pb_FileUtil.WriteFile(string.Format("{0}/{1}.mtl", path, name.Replace(" ", "_")), mat);
+					}
+					catch(System.Exception e)
+					{
+						Debug.LogWarning(string.Format("Failed writing obj to path: {0}\n{1}", string.Format("{0}/{1}.obj", path, name), e.ToString()));
+					}
+				}
+			}
 
 			return path;
 		}
